// lib/api/vega-graphql/lib/market.ts
export interface markets_markets_fees {
  __typename: "Fees";
  /**
   * The factors used to calculate the different fees
   */
  factors: markets_markets_fees_factors;
}

export interface markets_markets_data_market {
  __typename: "Market";
  /**
   * Market ID
   */
  id: string;
}

export interface markets_markets_data {
  __typename: "MarketData";
  /**
   * market id of the associated mark price
   */
  market: markets_markets_data_market;
  /**
   * the highest price level on an order book for buy orders.
   */
  bestBidPrice: string;
  /**
   * the aggregated volume being bid at the best bid price.
   */
  bestBidVolume: string;
  /**
   * the lowest price level on an order book for offer orders.
   */
  bestOfferPrice: string;
  /**
   * the aggregated volume being offered at the best offer price.
   */
  bestOfferVolume: string;
  /**
   * what state the market is in (auction, continuous etc)
   */
  marketTradingMode: MarketTradingMode;
  /**
   * the mark price (actually an unsgined int)
   */
  markPrice: string;
  /**
   * the sum of the size of all positions greater than 0.
   */
  openInterest: string;
  /**
   * RFC3339Nano time at which the next auction will start (null if none is scheduled)
   */
  auctionStart: string | null;
  /**
   * RFC3339Nano time at which the auction will stop (null if not in auction mode)
   */
  auctionEnd: string | null;
}
// ...
export interface markets_markets_tradableInstrument {
  __typename: "TradableInstrument";
  /**
   * An instance of or reference to a fully specified instrument.
   */
  instrument: markets_markets_tradableInstrument_instrument;
}

export interface markets_markets {
  __typename: "Market";
  /**
   * Market ID
   */
  id: string;
  /**
   * Market full name
   */
  name: string;
  /**
   * decimalPlaces indicates the number of decimal places that an integer must be shifted by in order to get a correct
   * number denominated in the currency of the Market. (uint64)
   * 
   * Examples:
   * Currency     Balance  decimalPlaces  Real Balance
   * GBP              100              0       GBP 100
   * GBP              100              2       GBP   1.00
   * GBP              100              4       GBP   0.01
   * GBP                1              4       GBP   0.0001   (  0.01p  )
   * 
   * GBX (pence)      100              0       GBP   1.00     (100p     )
   * GBX (pence)      100              2       GBP   0.01     (  1p     )
   * GBX (pence)      100              4       GBP   0.0001   (  0.01p  )
   * GBX (pence)        1              4       GBP   0.000001 (  0.0001p)
   */
  decimalPlaces: number;
  /**
   * Current state of the market
   */
  state: MarketState;
  /**
   * Fees related data
   */
  fees: markets_markets_fees;
  /**
   * marketData for the given market
   */
  data: markets_markets_data;
  /**
   * An instance of or reference to a tradable instrument.
   */
  tradableInstrument: markets_markets_tradableInstrument;
}

export interface markets {
  /**
   * One or more instruments that are trading on the VEGA network
   */
  markets: markets_markets[] | null;
}